feat: ajoute un avertissement lorsque la catégorie juridique de l'entreprise sélectionnée ne correspond pas au simulateur
